+++ examples/pipe.js
@@ -1,9 +1,9 @@
 "use strict";
 
-const { compose } = require("../build");
+const { pipe } = require("../build");
 
 const add = (x, y) => Promise.resolve(x + y);
 const square = (x) => Promise.resolve(x * x);
-const addSquare = compose([square, add]);
+const addSquare = pipe([add, square]);
 
 addSquare(1, 2).then(console.log); // 9