+++ examples/pipe.ts
@@ -1,8 +1,8 @@
 // tslint:disable: no-console
-import { compose } from "../lib";
+import { pipe } from "../lib";
 
 const add = (x: number, y: number) => Promise.resolve(x + y);
 const square = (x: number) => Promise.resolve(x * x);
-const addSquare = compose([square, add]);
+const addSquare = pipe([add, square]);
 
 addSquare(1, 2).then(console.log); // 9