+++ lib/operations/index.ts
@@ -1,6 +1,7 @@
 export * from "./compose";
 export * from "./delay";
 export * from "./map";
+export * from "./pipe";
 export * from "./promisify";
 export * from "./retry";
 export * from "./timeout";